Split packet_manager_test into simple_query_test and prepare_stmt_test
