+++ test/wire/simple_query_test.cpp
@@ -2,9 +2,9 @@
 //
 //                         Peloton
 //
-// packet_manager_test.cpp
+// simple_query_test.cpp
 //
-// Identification: test/wire/packet_manager_test.cpp
+// Identification: test/wire/simple_query_test.cpp
 //
 // Copyright (c) 2016-17, Carnegie Mellon University Database Group
 //
@@ -23,10 +23,10 @@ namespace peloton {
 namespace test {
 
 //===--------------------------------------------------------------------===//
-// Packet Manager Tests
+// Simple Query Tests
 //===--------------------------------------------------------------------===//
 
-class PacketManagerTests : public PelotonTest {};
+class SimpleQueryTests : public PelotonTest {};
 
 static void *LaunchServer(peloton::wire::LibeventServer libeventserver,int port) {
   try {
@@ -74,51 +74,6 @@ void *SimpleQueryTest(int port) {
   }
 
   LOG_INFO("[SimpleQueryTest] Client has closed");
-  return NULL;
-}
-
-/**
- * named prepare statement without parameters
- * TODO: add prepare's parameters when parser team fix the bug
- */
-
-void *PrepareStatementTest(int port) {
-  try {
-    pqxx::connection C(StringUtil::Format(
-            "host=127.0.0.1 port=%d user=postgres sslmode=disable",port));
-    LOG_INFO("[PrepareStatementTest] Connected to %s", C.dbname());
-    pqxx::work W(C);
-    
-    peloton::wire::LibeventSocket *conn =
-        peloton::wire::LibeventServer::GetConn(
-            peloton::wire::LibeventServer::recent_connfd);
-
-    // create table and insert some data
-    W.exec("DROP TABLE IF EXISTS employee;");
-    W.exec("CREATE TABLE employee(id INT, name VARCHAR(100));");
-    W.commit();
-    
-    pqxx::work W1(C);
-    W1.exec("INSERT INTO employee VALUES (1, 'Han LI');");
-    W1.exec("INSERT INTO employee VALUES (2, 'Shaokun ZOU');");
-    W1.exec("INSERT INTO employee VALUES (3, 'Yilei CHU');");
-
-    // test prepare statement
-    C.prepare("searchstmt","SELECT name FROM employee WHERE id=$1;");
-    // invocation as in variable binding
-    pqxx::result R = W1.prepared("searchstmt")(1).exec();
-    W1.commit();
-
-    // test prepared statement already in statement cache
-    // LOG_INFO("[Prepare statement cache] %d",conn->pkt_manager.ExistCachedStatement("searchstmt"));
-    EXPECT_TRUE(conn->pkt_manager.ExistCachedStatement("searchstmt"));
-
-    LOG_INFO("Prepare statement search result:%lu",R.size());
-  } catch (const std::exception &e) {
-    LOG_INFO("[PrepareStatementTest] Exception occurred");
-  }
-
-  LOG_INFO("[PrepareStatementTest] Client has closed");
   return NULL;
 }
 
@@ -196,7 +151,7 @@ TEST_F(PacketManagerTests, RollbackTest) {
  * Simple query test to guarantee both sides run correctly
  * Callback method to close server after client finishes
  */
-TEST_F(PacketManagerTests, SimpleQueryTest) {
+TEST_F(SimpleQueryTests, SimpleQueryTest) {
   peloton::PelotonInit::Initialize();
   LOG_INFO("Server initialized");
   peloton::wire::LibeventServer libeventserver;
@@ -209,24 +164,6 @@ TEST_F(PacketManagerTests, SimpleQueryTest) {
 
   // server & client running correctly
   SimpleQueryTest(port);
-
-  libeventserver.CloseServer();
-  serverThread.join();
-  peloton::PelotonInit::Shutdown();
-  LOG_INFO("Peloton has shut down");
-}
-
-TEST_F(PacketManagerTests, PrepareStatementTest) {
-  peloton::PelotonInit::Initialize();
-  LOG_INFO("Server initialized");
-  peloton::wire::LibeventServer libeventserver;
-  int port = 15721;
-  std::thread serverThread(LaunchServer, libeventserver,port);
-  while (!libeventserver.GetIsStarted()) {
-    sleep(1);
-  }
-
-  PrepareStatementTest(port);
 
   libeventserver.CloseServer();
   serverThread.join();
